x/nftvault/types/message_local_execution_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"nftvault/testutil/sample"
)

func TestMsgLocalExecution_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgLocalExecution
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgLocalExecution{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgLocalExecution{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
